feat:버그 수정 및 최적화, @property 정리

+++ nn/dropout.py
@@ -10,7 +10,7 @@ class Dropout(nn.Module):
         
     def forward(self, x:Tensor)->Tensor:
         if self.training and self.p > 0:
-            mask = torch.rand_like(x) > self.p
+            mask = torch.rand_like(x, dtype=torch.float16) > self.p
             x = x*mask/(1-self.p)
         return x
